requirements.txt
asgiref==3.7.2
asttokens==2.4.1
colorama==0.4.6
decorator==5.1.1
Django==4.2.7
djangorestframework==3.14.0
executing==2.0.1
iniconfig==2.0.0
ipython==8.17.2
jedi==0.19.1
matplotlib-inline==0.1.6
packaging==23.2
parso==0.8.3
pluggy==1.3.0
prompt-toolkit==3.0.41
pure-eval==0.2.2
Pygments==2.17.2
pytest==7.4.3
pytest-django==4.7.0
pytest-testdox==3.1.0
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.13.0
tzdata==2023.3
wcwidth==0.2.12
